pyproject.toml
[project]
name = "agents_from_scratch"
version = "0.1.0"
description = "Build an e-mail assistant from scratch"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "langsmith[pytest]>=0.4.37",
    "pandas",
    "matplotlib",
    "pytest",
    "pytest-xdist",
    "jupyter>=1.0.0",
    "langgraph-cli[inmem]>=0.4.0",
    "google-api-python-client>=2.128.0",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "python-dotenv",
    "pyppeteer",
    "html2text",
    "rich",
    "fastapi>=0.121.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "uvicorn[standard]>=0.34.3",
    "pydantic-settings>=2.11.0",
    "msal>=1.34.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ai_email_coach"]

[tool.setuptools.package-dir]
"ai_email_coach" = "src/ai_email_coach"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
